+++ resources/src/index.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import './index.css';
+import {App} from './App';
+
+// Learn more about TypeScript + React:
+// https://2ality.com/2018/04/type-notation-typescript.html
+// https://github.com/typescript-cheatsheets/react#section-1-setup-typescript-with-react
+// https://create-react-app.dev/docs/adding-typescript/
+
+ReactDOM.render(
+    <React.StrictMode>
+        <App/>
+    </React.StrictMode>,
+    document.getElementById('root')
+);